+++ storage/psql/db/migrations/000003_create_orders_table.up.sql
@@ -2,7 +2,7 @@ BEGIN;
 CREATE TABLE orders(
     order_id varchar(255) PRIMARY KEY,
     user_id integer REFERENCES users ON DELETE CASCADE,
-    accrual integer NOT NULL DEFAULT 0,
+    accrual decimal NOT NULL DEFAULT 0,
     status integer DEFAULT 1,
     uploaded_at timestamp
 );
